add trade stream to FTX
- refactored FTX to add FTX.US
- added ability to parse local DateTime in addition to UTC
- renamed Iso8601 enum to Iso8601UTC
