+++ src/ExchangeSharp/ExchangeSharp.csproj
@@ -27,7 +27,7 @@
   <ItemGroup>
     <None Include="../../icon.png" Pack="true" PackagePath="\" />
     <None Include="../../LICENSE.txt" Link="LICENSE.txt" Pack="true" PackagePath="" />
-		<None Include="../../README.md" Pack="true" PackagePath="\"/>
+		<None Include="../../README.md" Pack="true" PackagePath="\" />
   </ItemGroup>
 
   <ItemGroup>
